벼울)오잉

+++ BOHEMlogin/src/kdb/bohem/controller/DeleteAction.java
@@ -1,6 +1,7 @@
 package kdb.bohem.controller;
 
 import java.io.IOException;
+import java.io.PrintWriter;
 import java.sql.SQLException;
 import java.util.List;
 
@@ -16,32 +17,43 @@ public class DeleteAction implements Action {
 
 	@Override
 	public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		request.setCharacterEncoding("UTF-8");
+		
 		int comentIndex = Integer.parseInt(request.getParameter("comentIndex"));
 		String recipeIndex = request.getParameter("recipeIndex");
+		int userIndex = Integer.parseInt(request.getParameter("userIndex"));
 		//System.out.println(comentIndex);
 		//System.out.println("왔다");
 		String url ="errorView/error.jsp";
+		Comment comm=null;
+		//PrintWriter out = response.getWriter();
 		
 		try {
 			//System.out.println("왔니?"+comentIndex);
 			
-			Comment comm = CommentService.selectByComentIndex(comentIndex);
+			comm = CommentService.selectByComentIndex(comentIndex);
+			
 			System.out.println(comm);
 		} catch (SQLException e1) {
 			e1.printStackTrace();
 		}
 		
+		
 		try{
-			if(CommentService.delete(comentIndex) > 0){//삭제성공
-				List<Comment> comlist = CommentService.selectAll();
-				request.setAttribute("comlist", comlist);
-				url="comm?command=recipeindexselect&index="+recipeIndex;
-				
-				//response.sendRedirect("view/list.jsp");
-				// return;//메소드 빠져나가라.
-			 }else{//삭제실패
-				 throw new Exception("삭제되지 않았습니다.");
-			 }
+			if(comm.getUserIndex()==userIndex) {
+				if(CommentService.delete(comentIndex) > 0){//삭제성공
+					List<Comment> comlist = CommentService.selectAll();
+					request.setAttribute("comlist", comlist);
+					url="comm?command=recipeindexselect&index="+recipeIndex;
+					
+					//response.sendRedirect("view/list.jsp");
+					// return;//메소드 빠져나가라.
+				 }else{//삭제실패
+					 throw new Exception("삭제되지 않았습니다.");
+				 }
+			} else {
+					throw new Exception("회원 달라서");
+			}
 		}catch(Exception e){
 			e.printStackTrace();
 			request.setAttribute("errorMsg", e.getMessage());
